+++ src/commands/impls/server.rs
@@ -21,9 +21,19 @@ pub async fn quit<C: ClientLike>(client: &C) -> Result<(), RedisError> {
   let inner = client.inner().clone();
   _debug!(inner, "Closing Redis connection with Quit command.");
 
+  let (tx, rx) = oneshot_channel();
+  let command: RedisCommand = if inner.config.server.is_clustered() {
+    let response = ResponseKind::new_buffer(tx);
+    (RedisCommandKind::Quit, vec![], response).into()
+  } else {
+    let response = ResponseKind::Respond(Some(tx));
+    (RedisCommandKind::Quit, vec![], response).into()
+  };
   utils::set_client_state(&inner.state, ClientState::Disconnecting);
   inner.notifications.broadcast_close();
-  let _ = utils::request_response(client, || Ok((RedisCommandKind::Quit, vec![]))).await;
+
+  let _ = client.send_command(command)?;
+  let _ = rx.await??;
   utils::set_client_state(&inner.state, ClientState::Disconnected);
   Ok(())
 }
@@ -31,20 +41,25 @@ pub async fn quit<C: ClientLike>(client: &C) -> Result<(), RedisError> {
 pub async fn shutdown<C: ClientLike>(client: &C, flags: Option<ShutdownFlags>) -> Result<(), RedisError> {
   let inner = client.inner().clone();
   _debug!(inner, "Shutting down server.");
+
+  let args = if let Some(flags) = flags {
+    vec![flags.to_str().into()]
+  } else {
+    Vec::new()
+  };
+  let (tx, rx) = oneshot_channel();
+  let command: RedisCommand = if inner.config.server.is_clustered() {
+    let response = ResponseKind::new_buffer(tx);
+    (RedisCommandKind::Shutdown, args, response).into()
+  } else {
+    let response = ResponseKind::Respond(Some(tx));
+    (RedisCommandKind::Shutdown, args, response).into()
+  };
+  utils::set_client_state(&inner.state, ClientState::Disconnecting);
   inner.notifications.broadcast_close();
 
-  utils::set_client_state(&inner.state, ClientState::Disconnecting);
-  let _ = utils::request_response(client, move || {
-    let args = if let Some(flags) = flags {
-      vec![flags.to_str().into()]
-    } else {
-      Vec::new()
-    };
-
-    Ok((RedisCommandKind::Shutdown, args))
-  })
-  .await?;
-
+  let _ = client.send_command(command)?;
+  let _ = rx.await??;
   utils::set_client_state(&inner.state, ClientState::Disconnected);
   Ok(())
 }
@@ -105,7 +120,7 @@ pub async fn flushall_cluster<C: ClientLike>(client: &C) -> Result<(), RedisErro
   }
 
   let (tx, rx) = oneshot_channel();
-  let response = ResponseKind::new_buffer(client.inner().num_cluster_nodes(), tx);
+  let response = ResponseKind::new_buffer(tx);
   let command: RedisCommand = (RedisCommandKind::_FlushAllCluster, vec![], response).into();
   let _ = client.send_command(command)?;
 
@@ -151,7 +166,7 @@ pub async fn hello<C: ClientLike>(
   if client.inner().config.server.is_clustered() {
     let (tx, rx) = oneshot_channel();
     let mut command: RedisCommand = RedisCommandKind::_HelloAllCluster(version).into();
-    command.response = ResponseKind::new_buffer(client.inner().num_cluster_nodes(), tx);
+    command.response = ResponseKind::new_buffer(tx);
 
     let _ = client.send_command(command)?;
     let _ = rx.await??;
@@ -172,7 +187,7 @@ pub async fn auth<C: ClientLike>(client: &C, username: Option<String>, password:
 
   if client.inner().config.server.is_clustered() {
     let (tx, rx) = oneshot_channel();
-    let response = ResponseKind::new_buffer(client.inner().num_cluster_nodes(), tx);
+    let response = ResponseKind::new_buffer(tx);
     let command: RedisCommand = (RedisCommandKind::_AuthAllCluster, args, response).into();
     let _ = client.send_command(command)?;