+++ src/modules/inner.rs
@@ -36,7 +36,7 @@ const DEFAULT_NOTIFICATION_CAPACITY: usize = 32;
 
 #[cfg(feature = "metrics")]
 use crate::modules::metrics::MovingStats;
-use crate::protocol::types::Server;
+use crate::protocol::types::{Resolve, Server};
 
 pub type CommandSender = UnboundedSender<MultiplexerCommand>;
 pub type CommandReceiver = UnboundedReceiver<MultiplexerCommand>;
@@ -48,7 +48,7 @@ pub struct Notifications {
   /// A broadcast channel for the `on_error` interface.
   pub errors:         BroadcastSender<RedisError>,
   /// A broadcast channel for the `on_message` interface.
-  pub pubsub:         BroadcastSender<(String, RedisValue)>,
+  pub pubsub:         BroadcastSender<Message>,
   /// A broadcast channel for the `on_keyspace_event` interface.
   pub keyspace:       BroadcastSender<KeyspaceEvent>,
   /// A broadcast channel for the `on_reconnect` interface.
@@ -91,8 +91,8 @@ impl Notifications {
     }
   }
 
-  pub fn broadcast_pubsub(&self, channel: String, message: RedisValue) {
-    if let Err(_) = self.pubsub.send((channel, message)) {
+  pub fn broadcast_pubsub(&self, message: Message) {
+    if let Err(_) = self.pubsub.send(message) {
       debug!("{}: No `on_message` listeners.", self.id);
     }
   }
@@ -371,6 +371,11 @@ impl ServerState {
   }
 }
 
+// TODO make a config option for other defaults and extend this
+fn create_resolver(id: &ArcStr) -> Arc<dyn Resolve> {
+  Arc::new(DefaultResolver::new(id))
+}
+
 pub struct RedisClientInner {
   /// The client ID used for logging and the default `CLIENT SETNAME` value.
   pub id:            ArcStr,
@@ -393,8 +398,7 @@ pub struct RedisClientInner {
   /// Shared counters.
   pub counters:      ClientCounters,
   /// The DNS resolver to use when establishing new connections.
-  // TODO make this generic via the Resolve trait
-  pub resolver: DefaultResolver,
+  pub resolver:      AsyncRwLock<Arc<dyn Resolve>>,
   /// A backchannel that can be used to control the multiplexer connections even while the connections are blocked.
   pub backchannel:   Arc<AsyncRwLock<Backchannel>>,
   /// Server state cache for various deployment types.
@@ -417,7 +421,7 @@ pub struct RedisClientInner {
 impl RedisClientInner {
   pub fn new(config: RedisConfig, perf: PerformanceConfig, policy: Option<ReconnectPolicy>) -> Arc<RedisClientInner> {
     let id = ArcStr::from(format!("fred-{}", utils::random_string(10)));
-    let resolver = DefaultResolver::new(&id);
+    let resolver = AsyncRwLock::new(create_resolver(&id));
     let (command_tx, command_rx) = unbounded_channel();
     let notifications = Notifications::new(&id);
     let (config, policy) = (Arc::new(config), RwLock::new(policy));
@@ -473,6 +477,15 @@ impl RedisClientInner {
     if log_enabled!(level) {
       func(self.id.as_str())
     }
+  }
+
+  pub async fn set_resolver(&self, resolver: Arc<dyn Resolve>) {
+    let mut guard = self.resolver.write().await;
+    *guard = resolver;
+  }
+
+  pub async fn get_resolver(&self) -> Arc<dyn Resolve> {
+    self.resolver.read().await.clone()
   }
 
   pub fn client_name(&self) -> &str {