+++ src/multiplexer/clustered.rs
@@ -18,6 +18,7 @@ use arcstr::ArcStr;
 use futures::TryStreamExt;
 use std::{
   collections::{BTreeSet, HashMap},
+  iter::repeat,
   sync::Arc,
 };
 use tokio::task::JoinHandle;
@@ -83,9 +84,26 @@ pub async fn send_command(
 pub async fn send_all_cluster_command(
   inner: &Arc<RedisClientInner>,
   writers: &mut HashMap<Server, RedisWriter>,
-  command: RedisCommand,
+  mut command: RedisCommand,
 ) -> Result<(), RedisError> {
   let num_nodes = writers.len();
+  if let ResponseKind::Buffer {
+    ref mut frames,
+    ref mut expected,
+    ..
+  } = command.response
+  {
+    *expected = num_nodes;
+
+    _trace!(
+      inner,
+      "Allocating {} null responses in buffer for {}.",
+      num_nodes,
+      command.kind.to_str_debug(),
+    );
+    let mut guard = frames.lock();
+    *guard = repeat(Resp3Frame::Null).take(num_nodes).collect();
+  }
   let mut responder = match command.response.duplicate() {
     Some(resp) => resp,
     None => {
@@ -97,7 +115,13 @@ pub async fn send_all_cluster_command(
   };
 
   for (idx, (server, writer)) in writers.iter_mut().enumerate() {
-    _debug!(inner, "Sending all cluster command to {}", server);
+    _debug!(
+      inner,
+      "Sending all cluster command to {} with index {}, ID: {}",
+      server,
+      idx,
+      command.debug_id()
+    );
     let mut cmd_responder = responder.duplicate().unwrap_or(ResponseKind::Skip);
     cmd_responder.set_expected_index(idx);
     let mut cmd = command.duplicate(cmd_responder);
@@ -253,7 +277,7 @@ fn process_cluster_error(
     },
   };
 
-  if let Some(tx) = command.multiplexer_tx.take() {
+  if let Some(tx) = command.take_multiplexer_tx() {
     let response = match kind {
       ClusterErrorKind::Ask => MultiplexerResponse::Ask((slot, server, command)),
       ClusterErrorKind::Moved => MultiplexerResponse::Moved((slot, server, command)),
@@ -347,6 +371,12 @@ pub async fn process_response_frame(
       command
     }
   };
+  _trace!(
+    inner,
+    "Checking response to {} ({})",
+    command.kind.to_str_debug(),
+    command.debug_id()
+  );
   counters.decr_in_flight();
   responses::check_and_set_unblocked_flag(inner, &command).await;
 
@@ -527,6 +557,12 @@ pub async fn cluster_slots_backchannel(
     (protocol_utils::frame_to_results_raw(frame)?, host)
   };
   _trace!(inner, "Recv CLUSTER SLOTS response: {:?}", response);
+  if response.is_null() {
+    return Err(RedisError::new(
+      RedisErrorKind::Protocol,
+      "Invalid or missing CLUSTER SLOTS response.",
+    ));
+  }
 
   let mut new_cache = ClusterRouting::new();
   _debug!(inner, "Rebuilding cluster state from host: {}", host);