+++ src/multiplexer/commands.rs
@@ -7,14 +7,11 @@ use crate::{
   utils as client_utils,
 };
 use redis_protocol::resp3::types::Frame as Resp3Frame;
-use std::sync::Arc;
+use std::{sync::Arc, time::Duration};
+use tokio::time::sleep;
 
 #[cfg(feature = "mocks")]
 use crate::{modules::mocks::Mocks, protocol::utils as protocol_utils};
-#[cfg(feature = "mocks")]
-use std::time::Duration;
-#[cfg(feature = "mocks")]
-use tokio::time::sleep;
 #[cfg(feature = "partial-tracing")]
 use tracing_futures::Instrument;
 
@@ -151,6 +148,18 @@ async fn write_with_backpressure(
         _trace!(inner, "Ignore `Written` response.");
         break;
       },
+      Ok(Written::SentAll) => {
+        _trace!(inner, "Sent command to all servers.");
+        let _ = multiplexer.check_and_flush().await;
+        if let Some(mut command) = handle_multiplexer_response(inner, multiplexer, rx).await? {
+          // commands that are sent to all nodes are not retried after a connection closing
+          _warn!(inner, "Responding with canceled error after all nodes command failure.");
+          command.respond_to_caller(Err(RedisError::new_canceled()));
+          break;
+        } else {
+          break;
+        }
+      },
       Ok(Written::Sent((server, flushed))) => {
         _trace!(inner, "Sent command to {}. Flushed: {}", server, flushed);
         if is_blocking {
@@ -215,7 +224,8 @@ async fn process_pipeline(
     command.can_pipeline = true;
     command.skip_backpressure = true;
 
-    if let Err(e) = write_with_backpressure_t(inner, multiplexer, command, true).await {
+    let force_pipeline = !command.kind.is_all_cluster_nodes();
+    if let Err(e) = write_with_backpressure_t(inner, multiplexer, command, force_pipeline).await {
       // if the command cannot be written it will be queued to run later.
       // if a connection is dropped due to an error the reader will send a command to reconnect and retry later.
       _debug!(inner, "Error writing command in pipeline: {:?}", e);
@@ -434,6 +444,7 @@ async fn process_commands(inner: &Arc<RedisClientInner>, rx: &mut CommandReceive
 
 #[cfg(feature = "mocks")]
 pub async fn start(inner: &Arc<RedisClientInner>) -> Result<(), RedisError> {
+  sleep(Duration::from_millis(10)).await;
   if !client_utils::check_and_set_client_state(&inner.state, ClientState::Disconnected, ClientState::Connecting) {
     return Err(RedisError::new(
       RedisErrorKind::Unknown,
@@ -452,7 +463,6 @@ pub async fn start(inner: &Arc<RedisClientInner>) -> Result<(), RedisError> {
     },
   };
 
-  sleep(Duration::from_millis(10)).await;
   inner.notifications.broadcast_connect(Ok(()));
   inner.notifications.broadcast_reconnect();
   let result = process_commands(inner, &mut rx).await;
@@ -521,6 +531,7 @@ async fn process_commands(
 /// Start the command processing stream, initiating new connections in the process.
 #[cfg(not(feature = "mocks"))]
 pub async fn start(inner: &Arc<RedisClientInner>) -> Result<(), RedisError> {
+  sleep(Duration::from_millis(10)).await;
   if !client_utils::check_and_set_client_state(&inner.state, ClientState::Disconnected, ClientState::Connecting) {
     return Err(RedisError::new(
       RedisErrorKind::Unknown,