+++ CHANGELOG.md
@@ -3,17 +3,22 @@
 * Refactored the connection and protocol layer.
 * Add a manual `Pipeline` interface.
 * Rework the `Transaction` interface to buffer commands in memory before EXEC/DISCARD.
-* Switch from `Arc<String>` to `ArcStr` for shared server identifiers.
 * Rework the cluster discovery and failover implementation. 
 * Rework the MOVED/ASK implementation to more quickly and reliably follow cluster redirects.
 * Rework the sentinel interface to more reliably handle failover scenarios.
+* Fix several bugs related to detecting closed connections.
 * Support the `functions` interface.
 * Add some missing commands.
+* Add `Script`, `Library`, and `Function` structs. 
+* Add `Message` and `MessageKind` pubsub structs. 
+* Add a DNS configuration interface.
 * Rework the `native-tls` interface to support fully customizable TLS configurations.
 * Add `rustls` support.
   * Note: both TLS feature flags can be used at the same time.
+* Add an interface for mapping IP addresses back to hostnames for use cases that connect to clusters over TLS.
 * Add CI tests for Redis v7.
-* Add CI tests for clusters with TLS. 
+* Add CI tests for clusters with TLS.
+* Add a mocking layer interface.
 
 ### Updating from 5.x
 
@@ -26,9 +31,10 @@ New or breaking changes in 6.x:
 * Changed the type and fields on `BackpressurePolicy::Sleep`.
 * New [custom command interface](examples/custom.rs) for managing cluster hash slots.
 * Removed or renamed some fields on `RedisConfig`.
+* Changed the pubsub receiver interface to use `Message`.
 * Changed the `on_*` family of functions to return a [BroadcastReceiver](https://docs.rs/tokio/latest/tokio/sync/broadcast/struct.Receiver.html).
   * This usually means changing `next()` to `recv()` in `while let` loops, etc.
-* The `FromRedis` trait converts `RedisValue::Null` to `"nil"`.
+* The `FromRedis` trait converts `RedisValue::Null` to `"nil"` with `String` and `Str`.
 
 ## 5.2.0